Make the arrows able to be told about the range of what they're marking.
Arrows should be red if you're almost certainly standing in something
now, orange if you might be, yellow if you're probably safe and green if
you're guranteed to be safe.  The ranges are tightened up so you don't
need to run as far now to get a green arrow.

Blood Prince Council (23): Ajust the range for the Shock Vortex arrow.

Lich King (67): Adjust the range for the Shadow Trap arrow.

Putricide (41): Adjust the range for the Maleable Goo arrow.

This is impelemented by adding range1, range2 and range3 options to the
arrow definition.  If ommitted they default to range1 = 10, range2 =
range1*2, range3 = range1*3.  range1 should always be the shortest
distance and increasing out to range3.  For AWAY arrows, range1 should
be the size of the effect of what you're running away from.  For TOWARD
arrows range1 should be the maximum distance to be considered at the
object.